Post ui refactor
- added file in posts folder to hold post ui actions
