+++ assets/scripts/comments/events.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const ui = require('../ui')
+const ui = require('./ui')
 const api = require('./api')
 const store = require('../store')
 const getFormFields = require('../../../lib/get-form-fields')
@@ -54,7 +54,6 @@ function confirmDeleteComment (event) {
     postId: store.viewItemId }
   }
   $('.main-comment-buttons').removeClass('hidden')
-  $(`.delete-check[data-id=${store.deleteCommentId}]`).addClass('hidden')
   api.deleteComment(data, store.deleteCommentId)
     .then(ui.confirmDeleteCommentSuccess)
     .catch(ui.deleteCommentsFailure)
