+++ assets/scripts/posts/events.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const ui = require('../ui')
+const ui = require('./ui')
 const api = require('./api')
 const store = require('../store')
 const getFormFields = require('../../../lib/get-form-fields')
