+++ assets/scripts/ui.js
@@ -1,74 +1,7 @@
 'use strict'
 
 const store = require('./store')
-const showPostsTemplate = require('./templates/blog-post.handlebars')
-const showUserPostsTemplate = require('./templates/owner-blog-posts.handlebars')
-const populateUpdateTemplate = require('./templates/populate-update-template.handlebars')
-const postApi = require('./posts/api')
 const api = require('./auth/api')
-
-// POSTS CRUD
-
-function viewPostsSuccess (data) {
-  data.posts = data.posts.sort(function (a, b) {
-    a = new Date(a.createdAt)
-    b = new Date(b.createdAt)
-    return a > b ? -1 : a < b ? 1 : 0
-  })
-  data.posts.forEach(x => { x.createdAt = (new Date(x.createdAt).toDateString()) })
-  const showPostsHtml = showPostsTemplate({posts: data.posts})
-  $('.content').html(showPostsHtml)
-  $('#post-content').removeClass('hidden')
-  $('#create-post-menu').addClass('hidden')
-}
-
-function viewUserPostsSuccess (data) {
-  data.posts = data.posts.sort(function (a, b) {
-    a = new Date(a.createdAt)
-    b = new Date(b.createdAt)
-    return a > b ? -1 : a < b ? 1 : 0
-  })
-  data.posts.forEach(x => { x.createdAt = (new Date(x.createdAt).toDateString()) })
-  const showUserPostsHtml = showUserPostsTemplate({posts: data.posts})
-  $('.content').html(showUserPostsHtml)
-  $('#post-content').removeClass('hidden')
-  $('#create-post-menu').addClass('hidden')
-}
-
-function deletePostsSuccess () {
-  console.log('delete Posts success')
-  postApi.viewPosts()
-    .then(refreshListSuccess)
-    .catch(refreshListFailure)
-}
-
-function selectUpdatePostsSuccess (data) {
-  console.log('selectUpdatePostsSuccess')
-  const populateUpdateHtml = populateUpdateTemplate({post: data.post[0]})
-  $('#editModalLong').html(populateUpdateHtml)
-}
-
-function updatePostsSuccess () {
-  postApi.viewPosts()
-    .then(refreshListSuccess)
-    .catch(refreshListFailure)
-}
-
-function refreshListSuccess (data) {
-  const showUserPostsHtml = showUserPostsTemplate({posts: data.posts})
-  $('.content').html(showUserPostsHtml)
-}
-
-function createPostSuccess () {
-  console.log('create post success')
-  $('.message').text('Post created')
-}
-
-function cancelDeleteComment () {
-  $(`.main-comment-buttons`).removeClass('hidden')
-  $(`.delete-check[data-id=${store.deleteCommentId}]`).addClass('hidden')
-  $(`.delete-comment[data-id=${store.deleteCommentId}]`).removeClass('hidden')
-}
 
 // AUTHENTICATION
 
@@ -102,24 +35,6 @@ function showWritePost () {
   $('#create-post-menu').removeClass('hidden')
 }
 
-// FAILURES
-
-function deletePostsFailure () {
-  console.log('delete Posts failed')
-}
-
-function refreshListFailure () {
-  console.log('delete Posts failed')
-}
-
-function viewPostsFailure () {
-  console.log('viewPosts failed')
-}
-
-function createPostFailure () {
-  $('.messge').text('failed to create post')
-}
-
 function signUpFailure () {
   console.log('signUp failed')
 }
@@ -149,15 +64,5 @@ module.exports = {
   changePasswordFailure,
   signOutSuccess,
   signOutFailure,
-  viewPostsSuccess,
-  viewUserPostsSuccess,
-  viewPostsFailure,
-  createPostSuccess,
-  createPostFailure,
-  showWritePost,
-  deletePostsSuccess,
-  deletePostsFailure,
-  updatePostsSuccess,
-  selectUpdatePostsSuccess,
-  cancelDeleteComment
+  showWritePost
 }
